makes offchain batch deliver with sig work
