+++ packages/packages.json
@@ -37,19 +37,19 @@
         "custom/dvilela/gemini_prediction/0.1.0": "bafybeigvwflupxzbjgmaxcxml5vkez3obl4fjo6bxzhquq56urnviq32u4",
         "protocol/valory/acn_data_share/0.1.0": "bafybeih5ydonnvrwvy2ygfqgfabkr47s4yw3uqxztmwyfprulwfsoe7ipq",
         "protocol/valory/websocket_client/0.1.0": "bafybeifjk254sy65rna2k32kynzenutujwqndap2r222afvr3zezi27mx4",
-        "contract/valory/agent_mech/0.1.0": "bafybeib6xrew4kfvus376hqcrkt6e35yfr4jp7bwbrlb3eaav262e5ns4m",
+        "contract/valory/agent_mech/0.1.0": "bafybeifugr6fgvekv3g2zsu7xz2vteif3x7zr27tf3ofaeoqguycdv5eri",
         "contract/valory/agent_registry/0.1.0": "bafybeiarzhzs2wm2sl47qg37tqoc3qok54enxlcj6vx3hldozg537uslnq",
         "contract/valory/hash_checkpoint/0.1.0": "bafybeibpfllgkoa2fb47gtb7qwlb6gya5yufgm7m7u5f3cdwxuu2r2grw4",
-        "contract/valory/mech_marketplace/0.1.0": "bafybeiafl4nbxpzs45lipksegw76rg6iurrvsr7xkxdhjklrpkniofctfy",
+        "contract/valory/mech_marketplace/0.1.0": "bafybeiftw4xjshpevnx44hmglkflyxvsy47u3ui7x26rl5cvvspminov5a",
         "connection/valory/websocket_client/0.1.0": "bafybeic4ag3gqc7kd3k2o3pucddj2odck5yrfbgmwh5veqny7zao5qayli",
         "skill/valory/contract_subscription/0.1.0": "bafybeiefuemlp75obgpxrp6iuleb3hn6vcviwh5oetk5djbuprf4xsmgjy",
-        "skill/valory/mech_abci/0.1.0": "bafybeicfjwtjboqpn4usy4mlcxk32wjbhuy7jsbduzku64njtyauis5p4i",
-        "skill/valory/task_submission_abci/0.1.0": "bafybeiat4ghphonpgu5anym2jjt6rmf54csd4jiwsiyne3i2c2x34wsbs4",
-        "skill/valory/task_execution/0.1.0": "bafybeiawa7ufijo4jcvb4ftynwcnmqfrccwkkqpa4rlegt7srkowekqqqm",
+        "skill/valory/mech_abci/0.1.0": "bafybeihmahr7zqrgn7r5yojmscm7d5dthcsml6sntieciyeuhxiiupqwia",
+        "skill/valory/task_submission_abci/0.1.0": "bafybeifubbhcjeucggshk3sujeeerq4bsdhvt2d7lhs4vegk2h67owi6o4",
+        "skill/valory/task_execution/0.1.0": "bafybeihvvatateg3njw3gkgnnv5qkuaqgvz7i3aljs4vhonuvchlwbioam",
         "skill/valory/websocket_client/0.1.0": "bafybeif7rrvsu6z4evqkhblxj3u6wwv2eqou576hgkyoehxuj7cntw7o2m",
-        "skill/valory/subscription_abci/0.1.0": "bafybeias2n7742yl4csch2ciaqea2vlugvew4j7zbfrryhevbv4fck2hfe",
-        "agent/valory/mech/0.1.0": "bafybeife5qaddjtvgxx62a6tnataowaokf6ul4vaqa65foaxgbxsehmtsm",
-        "service/valory/mech/0.1.0": "bafybeidamhrjheh4vaj4maifp3rjy7i3ybki7eo7sa5eaobhxdzudovohi",
+        "skill/valory/subscription_abci/0.1.0": "bafybeidvtimvthvx4msd4e6kvgcizpbodgxd2pd26c5fpkwn22jhpd56a4",
+        "agent/valory/mech/0.1.0": "bafybeibcqlrwm56esj6kqkrj7bexdercebfh4vx7pptgy6fuslyt4fqut4",
+        "service/valory/mech/0.1.0": "bafybeihx44evnodcwcyajizhqghxc4fm73fqejn23fbdiqed2dfrtdmdk4",
         "service/valory/mech_quickstart/0.1.0": "bafybeic7nyt734ysd2agj352zlxhxxemi6laeevclvepzzlm53oqp3hwxi"
     },
     "third_party": {
