+++ packages/valory/contracts/agent_mech/contract.yaml
@@ -7,8 +7,8 @@ license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: bafybeigpq5lxfj2aza6ok3fjuywtdafelkbvoqwaits7regfbgu4oynmku
-  build/AgentMech.json: bafybeifbx2dovjm7ufoufvxwb5n3tyfcwysecaggpcds4caanqlpfg5dqm
-  contract.py: bafybeig5pnbegw4qw3vltgfaay5edf55bkeoewy5dqz6q7dsdclu2doxji
+  build/AgentMech.json: bafybeienigznrwv45y4f4dth4r2xvqtffrdphi3qwqta34fakhpzjlym3i
+  contract.py: bafybeihkcptglccdytdqzzp6vjqcoalt75ghz3plcbqw7zeg3w6bdbfrby
 fingerprint_ignore_patterns: []
 class_name: AgentMechContract
 contract_interface_paths:
