+++ packages/valory/contracts/mech_marketplace/contract.py
@@ -341,36 +341,3 @@ class MechMarketplaceContract(Contract):
             simulation_ok = False
 
         return dict(data=simulation_ok)
-
-
-    @classmethod
-    def get_offchain_deliver_data(
-        cls,
-        ledger_api: LedgerApi,
-        contract_address: str,
-        requester: str,
-        requestDatas: List[str],
-        signatures: List[str],
-        deliverDatas: List[str],
-        deliveryRates: List[int],
-        paymentData: str,
-    ) -> JSONLike:
-
-        ledger_api = cast(EthereumApi, ledger_api)
-        contract_instance = cls.get_instance(ledger_api, contract_address)
-        data = contract_instance.encodeABI(
-            fn_name="deliverMarketplaceWithSignatures",
-            args=[
-                requester,
-                requestDatas,
-                signatures,
-                deliverDatas,
-                deliveryRates,
-                paymentData,
-            ],
-        )
-
-        simulation_ok = cls.simulate_tx(
-            ledger_api, contract_address, contract_address, data
-        ).pop("data")
-        return {"data": bytes.fromhex(data[2:]), "simulation_ok": simulation_ok}  # type: ignore
