+++ packages/valory/services/mech/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeif7ia4jdlazy6745ke2k2x5yoqlwsgwr6sbztbgqtwvs3ndm2p7ba
 fingerprint_ignore_patterns: []
-agent: valory/mech:0.1.0:bafybeife5qaddjtvgxx62a6tnataowaokf6ul4vaqa65foaxgbxsehmtsm
+agent: valory/mech:0.1.0:bafybeibcqlrwm56esj6kqkrj7bexdercebfh4vx7pptgy6fuslyt4fqut4
 number_of_agents: 4
 deployment:
   agent:
